requirements.txt
aiofiles==23.2.1
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
Django==5.2.1
djangorestframework==3.16.0
docker==7.1.0
ecdsa==0.19.1
fastapi==0.104.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
Jinja2==3.1.2
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
pywin32==310
PyYAML==6.0.2
requests==2.31.0
rpds-py==0.24.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
sqlparse==0.5.3
starlette==0.27.0
tablib==3.8.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.24.0
watchfiles==1.0.5
websockets==15.0.1
whitenoise==6.9.0
windows-curses==2.4.1
xlrd==2.0.1
xlwt==1.3.0
